feat(run_harvest): Initial Harvest Job display
