+++ app/views/jobs/_jobs.html.erb
@@ -12,8 +12,8 @@
       </div>
 
       <div class="col-auto">
-        <% path = content_partner_extraction_definition_extraction_job_path(job.extraction_definition.content_partner, job.extraction_definition, job) %>
-        <% cancel_path = cancel_content_partner_extraction_definition_extraction_job_path(job.extraction_definition.content_partner, job.extraction_definition, job) %>
+        <% path = send("content_partner_#{type}_definition_#{type}_job_path", content_partner, definition, job) %>
+        <% cancel_path = send("cancel_content_partner_#{type}_definition_#{type}_job_path", content_partner, definition, job) %>
 
         <% if job.queued? || job.running? %>
           <%= button_to "Cancel job", cancel_path, class: "btn btn-outline-danger me-1", form_class: 'd-inline' %>
